Process logscale selection and layer min/max values
These are linked, as layer min determines whether logscale can be used.
Layer min/max (aka. layer range) is a state of AltMapController
(because this is lowest ancestor in which it is used). Props propagate
it down in to DataMap (which also uses it). Callbacks propagate it back
up the hierarchy, natch.